+++ example/lib/router.dart
@@ -2,6 +2,7 @@ import 'package:go_router/go_router.dart';
 
 import 'main.dart';
 import 'model.dart';
+import 'widgets.dart';
 
 final appRouter = GoRouter(routes: [
   GoRoute(path: '/', builder: (context, state) => const Home()),